+++ repos/gems/src/server/terminal/session.h
@@ -100,9 +100,9 @@ class Terminal::Session_component : public Rpc_object<Session, Session_component
 		{
 			char const *src = _io_buffer.local_addr<char const>();
 
-			unsigned const max = min(num_bytes, _io_buffer.size());
+			size_t const max = min(num_bytes, _io_buffer.size());
 
-			unsigned i = 0;
+			size_t i = 0;
 			for (Utf8_ptr utf8(src); utf8.complete() && i < max; ) {
 
 				_character_consumer.consume_character(
@@ -121,7 +121,7 @@ class Terminal::Session_component : public Rpc_object<Session, Session_component
 			/* we don't support UTF-8 sequences split into multiple writes */
 			if (i != num_bytes) {
 				warning("truncated UTF-8 sequence");
-				for (unsigned j = i; j < num_bytes; j++)
+				for (size_t j = i; j < num_bytes; j++)
 					warning("(unhandled value ", Hex(src[j]), ")");
 			}
